Move test_indexing.py
